move remaining filters to filters dir
